Added update availability and search functionalities.

+++ src/app/components/buttons/button-remove-book/button-remove-book.component.ts
@@ -1,17 +1,25 @@
 import { Component, Input } from '@angular/core';
 import { BookService } from '../../../services/book.service';
+import { MatDialog, MatDialogModule } from '@angular/material/dialog';
+import { ConfirmDialogComponent } from '../../modals/confirm-dialog/confirm-dialog.component';
+import { CommonModule } from '@angular/common';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-button-remove-book',
   standalone: true,
-  imports: [],
+  imports: [MatDialogModule, CommonModule],
   templateUrl: './button-remove-book.component.html',
   styleUrl: './button-remove-book.component.css',
 })
 export class ButtonRemoveBookComponent {
   @Input() isbn!: string;
 
-  constructor(private bookService: BookService) {}
+  constructor(
+    private bookService: BookService,
+    private dialog: MatDialog,
+    private router: Router
+  ) {}
 
   removeBook() {
     if (!this.isbn) {
@@ -19,13 +27,21 @@ export class ButtonRemoveBookComponent {
       return;
     }
 
-    this.bookService.deleteBook(this.isbn).subscribe({
-      next: () => {
-        console.log('Book removed successfully');
-      },
-      error: (err) => {
-        console.error('Failed to remove book: ', err);
-      },
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      width: '30vw',
+      disableClose: true,
+    });
+
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result) {
+        this.bookService.deleteBook(this.isbn).subscribe({
+          next: () => {
+            console.log('Livro removido com sucesso');
+            this.router.navigate(['/books/available']);
+          },
+          error: (err) => console.error('Erro ao remover livro: ', err),
+        });
+      }
     });
   }
 }
